Oh boy, gstreamer 0.8 changed the size of the volume element's volume property from float to double.  Of course, this couldn't be caught during compilation since gstreamer uses C, and the function in question is essential scanf() style, but at least I was able to set speed records recompiling this damn file 50 times.
Committing now because this is a memory corruptor bug (although I don't think it's a security flaw).

svn path=/trunk/kdemultimedia/juk/; revision=375809

+++ gstreamerplayer.cpp
@@ -72,9 +72,9 @@ void GStreamerPlayer::setVolume(float volume)
 
 float GStreamerPlayer::volume() const
 {
-    gfloat value;
+    gdouble value;
     g_object_get(G_OBJECT(m_volume), "volume", &value, 0);
-    return value;
+    return (float) value;
 }
 
 bool GStreamerPlayer::playing() const
